requirements.txt
# django-phonenumber-field-6.0.0
#django_phonenumber_field-6.0.0-py3-none-any.whl
django-phonenumber-field[phonenumbers]
requests==2.22.0
django==2.2.16
django-import-export
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
django-xml==3.0.0
csvImporter==0.1.3.2
django-adaptors==0.2.5
# Pillow==7.0.0